# Conflicts:
#	Gruntfile.js
#	bp-loader.php
#	src/bp-core/compatibility/bp-incompatible-plugins-helper.php
#	src/bp-loader.php
#	src/languages/buddyboss.pot
#	src/readme.txt

+++ src/bp-media/bp-media-functions.php
@@ -3327,15 +3327,23 @@ function bp_media_get_preview_image_url( $media_id, $attachment_id, $size = 'bb-
 		if ( wp_attachment_is_image( $attachment_id ) && file_exists( $output_file_src ) ) {
 			$media = new BP_Media( $media_id );
 
-			$upload_directory = wp_get_upload_dir();
-			$symlinks_path    = bp_media_symlink_path();
+			if ( bb_enable_symlinks() ) {
 
-			$preview_attachment_path = $symlinks_path . '/' . md5( $media_id . $attachment_id . $media->privacy . $size );
-			if ( ! file_exists( $preview_attachment_path ) && $generate ) {
-				bp_media_create_symlinks( $media, $size );
-			}
+				$upload_directory = wp_get_upload_dir();
+				$symlinks_path    = bp_media_symlink_path();
 
-			$attachment_url = bb_core_symlink_absolute_path( $preview_attachment_path, $upload_directory );
+				$preview_attachment_path = $symlinks_path . '/' . md5( $media_id . $attachment_id . $media->privacy . $size );
+				if ( ! file_exists( $preview_attachment_path ) && $generate ) {
+					bp_media_create_symlinks( $media, $size );
+				}
+
+				$attachment_url = bb_core_symlink_absolute_path( $preview_attachment_path, $upload_directory );
+
+			} else {
+				$media_id       = 'forbidden_' . $media_id;
+				$attachment_id  = 'forbidden_' . $attachment_id;
+				$attachment_url = trailingslashit( buddypress()->plugin_url ) . 'bp-templates/bp-nouveau/includes/media/preview.php?id=' . base64_encode( $attachment_id ) . '&id1=' . base64_encode( $media_id ) . '&size=' . $size;
+            }
 
 		}
 	}
@@ -3582,10 +3590,30 @@ function bb_media_user_can_access( $id, $type ) {
 		$activity_id    = $document->activity_id;
 	}
 
-	if ( $media_privacy === 'comment' && bp_is_active( 'activity' ) && ! empty( $activity_id )) {
-		$activity = new BP_Activity_Activity( $activity_id );
-		if( ! empty( $activity->id ) && ! empty( $activity->privacy ) ) {
-			$media_privacy  = $activity->privacy;
+	if ( 'comment' === $media_privacy && bp_is_active( 'activity' ) && ! empty( $activity_id ) ) {
+		$hierarchy = bb_get_activity_hierarchy( $activity_id );
+		if ( ! empty( $hierarchy ) ) {
+			$main_parent_id = end( $hierarchy );
+			if ( ! empty( $main_parent_id ) ) {
+				$parent_activity = new BP_Activity_Activity( $main_parent_id['id'] );
+				if ( ! empty( $parent_activity->id ) && ! empty( $parent_activity->privacy ) ) {
+					$media_privacy = $parent_activity->privacy;
+				}
+			}
+		}
+
+		if ( empty( $media_privacy ) ) {
+			$activity = new BP_Activity_Activity( $activity_id );
+			if ( ! empty( $activity->item_id ) ) {
+				$parent_activity = new BP_Activity_Activity( $activity->item_id );
+				if ( ! empty( $parent_activity->id ) && ! empty( $parent_activity->privacy ) ) {
+					$media_privacy = $parent_activity->privacy;
+				} else if ( ! empty( $activity->id ) && ! empty( $activity->privacy ) ) {
+					$media_privacy = $activity->privacy;
+				}
+			} else if ( ! empty( $activity->id ) && ! empty( $activity->privacy ) ) {
+				$media_privacy = $activity->privacy;
+			}
 		}
 	}
 
@@ -3783,6 +3811,10 @@ function bb_media_user_can_access( $id, $type ) {
  */
 function bb_media_delete_older_symlinks() {
 
+    if ( ! bb_enable_symlinks() ) {
+        return;
+    }
+
 	// Get documents previews symlink directory path.
 	$dir     = bp_media_symlink_path();
 	$max_age = 3600 * 24 * 1; // Delete the file older then 1 day.
@@ -3804,7 +3836,7 @@ function bb_media_delete_older_symlinks() {
 			continue;
 		}
 		$file = $dir . '/' . $file;
-		if ( filemtime( $file ) < $limit ) {
+		if ( file_exists( $file ) && filemtime( $file ) < $limit ) {
 			$list[] = $file;
 			unlink( $file );
 		}
@@ -3815,3 +3847,4 @@ function bb_media_delete_older_symlinks() {
 
 }
 bp_core_schedule_cron( 'bb_media_deleter_older_symlink', 'bb_media_delete_older_symlinks' );
+
